requirements.txt
# VisionGuard AI - Requirements
# Latest and easiest tech stack for ML (December 2025)

# ===== Core Deep Learning =====
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# ===== Transformers & Pretrained Models =====
transformers>=4.36.0          # Hugging Face - latest pretrained models
timm>=0.9.12                  # PyTorch Image Models - EfficientNet, ViT, etc.
accelerate>=0.25.0            # Fast model loading

# ===== Computer Vision =====
opencv-python>=4.8.1          # Image/video processing
Pillow>=10.1.0                # Image handling
albumentations>=1.3.1         # Advanced augmentations

# ===== Object Detection (YOLO) =====
ultralytics>=8.1.0            # YOLOv8 - latest and easiest
supervision>=0.17.0           # Detection utilities

# ===== Face Recognition =====
insightface>=0.7.3            # ArcFace - SOTA face recognition
facenet-pytorch>=2.5.3        # Alternative face embeddings
deepface>=0.0.79              # Easy face analysis framework

# ===== API & Web =====
fastapi>=0.108.0              # Modern async API
uvicorn[standard]>=0.25.0     # ASGI server
streamlit>=1.29.0             # Easy UI framework
python-multipart>=0.0.6       # File upload support

# ===== Authentication & Security =====
python-jose[cryptography]>=3.3.0  # JWT token handling
passlib[bcrypt]>=1.7.4        # Password hashing
python-dotenv>=1.0.0          # Environment variables (moved from utilities)

# ===== Database =====
pymongo>=4.6.0                # MongoDB driver
dnspython>=2.4.2              # Required for MongoDB+SRV connections

# ===== Telegram Bot =====
python-telegram-bot>=21.0     # Telegram Bot API (async)

# ===== AI APIs (Optional) =====
google-genai>=1.0.0           # Gemini API (new SDK — replaces google-generativeai)
google-generativeai>=0.3.0    # kept for deepfake fallback (legacy)

# ===== TTS =====
edge-tts>=6.1.9              # Microsoft neural TTS (free, online)

# ===== Utilities =====
numpy>=1.24.0
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.3.0
scipy>=1.11.0

# ===== RAG + Vector Search (Assistant) =====
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0

# ===== Offline TTS (Assistant voice) =====
pyttsx3>=2.90

# ===== Voice Lock (owner speaker verification) =====
resemblyzer>=0.1.2              # Pretrained GE2E speaker embeddings
soundfile>=0.12.1               # WAV decoding for resemblyzer

# ===== Configuration & Logging =====
pyyaml>=6.0.1
loguru>=0.7.2

# ===== Optional: Advanced Features =====
grad-cam>=1.4.8               # Visualization
onnx>=1.15.0                  # Model export
onnxruntime>=1.16.0           # Fast inference

# ===== Development Tools =====
pytest>=7.4.3
black>=23.12.0
flake8>=6.1.0
ipython>=8.18.0
jupyter>=1.0.0
